requirements.txt
Keras==2.2.4
matplotlib==3.5.2
numpy==1.20.3
pandas==1.4.3
Pillow==10.0.1
requests==2.28.1
scikit_learn==1.3.1
streamlit==1.12.2
tensorflow==2.9.1
